Genflow/backend/src/requirements.txt
fastapi
uvicorn
google-generativeai
pandas
numpy
sentence-transformers
scikit-learn
python-dotenv
pydantic-settings
matplotlib
pillow
requests
